humidity is kind of working but idfk

+++ humidity.h
@@ -14,7 +14,7 @@ uint16_t _humidity;
 enum dataByteLocations { HUMIDITY_H, HUMIDITY_L, TEMP_H, TEMP_L, CHECKSUM };
 
 void init_humidity();
-uint8_t update_humidity();
+uint64_t update_humidity();
 uint16_t get_humidity();
 _Bool checksum(char check, char * data, unsigned int datalen);
 uint8_t errorExit(int code);
